# docs/apis/swagger.json
{
  "swagger" : "2.0",
  "info" : {
...
    "version" : "1.1.0",
    "title" : "DMaaP Bus Controller REST API",
    "termsOfService" : "http://www.apache.org/licenses/LICENSE-2.0",
    "contact" : {
      "url" : "http://www.onap.org"
    },
    "license" : {
      "name" : "Licensed under the Apache License, Version 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "www.[host]:[port]",
  "basePath" : "/webapi",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/dmaap" : {
      "get" : {
        "tags" : [ "dmaap" ],
        "summary" : "return dmaap details",
        "description" : "returns the `dmaap` object, which contains system wide configuration settings",
        "operationId" : "getDmaap",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Dmaap"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "dmaap" ],
        "summary" : "return dmaap details",
        "description" : "Create a new DMaaP set system wide configuration settings for the *dcaeEnvironment*.  Deprecated with introduction of persistence in 1610.",
        "operationId" : "addDmaap",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Dmaap"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "dmaap" ],
        "summary" : "return dmaap details",
        "description" : "Update system settings for *dcaeEnvironment*.",
        "operationId" : "updateDmaap",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Dmaap"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
    "/feeds" : {
      "get" : {
        "tags" : [ "Feeds" ],
        "summary" : "return Feed details",
        "description" : "Returns array of  `Feed` objects.",
        "operationId" : "getFeeds",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "feedName",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "version",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "match",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Feed"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Feeds" ],
        "summary" : "return Feed details",
        "description" : "Create a of  `Feed` object.",
        "operationId" : "addFeed",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "useExisting",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Feed"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/feeds/sync" : {
      "put" : {
        "tags" : [ "Feeds" ],
        "summary" : "sync feeds to existing DR",
        "description" : "When Bus Controller is deployed after DR, then it is possiblethat DR has previous provisioning data that needs to be importedinto Bus Controller.",
        "operationId" : "syncFeeds",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "hard",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Feed"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/feeds/{id}" : {
      "get" : {
        "tags" : [ "Feeds" ],
        "summary" : "return Feed details",
        "description" : "Retrieve a  `Feed` object, specified by id.",
        "operationId" : "getFeed",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/DR_Pub"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Feeds" ],
        "summary" : "return Feed details",
        "description" : "Update a  `Feed` object, specified by id.",
        "operationId" : "updateFeed",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Feed"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Feeds" ],
        "summary" : "return Feed details",
        "description" : "Delete a  `Feed` object, specified by id.",
        "operationId" : "deleteFeed",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Feed"
            }
          },
          "204" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Feed"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ApiError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "fields" : {
          "type" : "string"
        },
        "is2xx" : {
          "type" : "boolean"
        }
      }
    },
...
    "DR_Pub" : {
      "type" : "object",
      "properties" : {
        "lastMod" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "datestamp for last update to this object"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "EMPTY", "NEW", "STAGED", "VALID", "INVALID", "DELETED" ]
        },
        "dcaeLocationName" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "userpwd" : {
          "type" : "string"
        },
        "feedId" : {
          "type" : "string"
        },
        "pubId" : {
          "type" : "string"
        },
        "feedName" : {
          "type" : "string"
        },
        "feedVersion" : {
          "type" : "string"
        }
      }
    },
    "DR_Sub" : {
      "type" : "object",
      "properties" : {
        "lastMod" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "datestamp for last update to this object"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "EMPTY", "NEW", "STAGED", "VALID", "INVALID", "DELETED" ]
        },
        "dcaeLocationName" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "userpwd" : {
          "type" : "string"
        },
        "feedId" : {
          "type" : "string"
        },
        "deliveryURL" : {
          "type" : "string"
        },
        "logURL" : {
          "type" : "string"
        },
        "subId" : {
          "type" : "string"
        },
        "use100" : {
          "type" : "boolean"
        },
        "suspended" : {
          "type" : "boolean"
        },
        "owner" : {
          "type" : "string"
        },
        "guaranteedDelivery" : {
          "type" : "boolean"
        },
        "guaranteedSequence" : {
          "type" : "boolean"
        },
        "privilegedSubscriber" : {
          "type" : "boolean"
        },
        "decompress" : {
          "type" : "boolean"
        },
        "feedName" : {
          "type" : "string"
        },
        "feedVersion" : {
          "type" : "string"
        }
      }
    },
...
    "Dmaap" : {
      "type" : "object",
      "properties" : {
        "lastMod" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "datestamp for last update to this object"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "EMPTY", "NEW", "STAGED", "VALID", "INVALID", "DELETED" ]
        },
        "version" : {
          "type" : "string"
        },
        "topicNsRoot" : {
          "type" : "string"
        },
        "dmaapName" : {
          "type" : "string"
        },
        "drProvUrl" : {
          "type" : "string"
        },
        "bridgeAdminTopic" : {
          "type" : "string"
        },
        "loggingUrl" : {
          "type" : "string"
        },
        "nodeKey" : {
          "type" : "string"
        },
        "accessKeyOwner" : {
          "type" : "string"
        }
      }
    },
    "Feed" : {
      "type" : "object",
      "properties" : {
        "lastMod" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "datestamp for last update to this object"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "EMPTY", "NEW", "STAGED", "VALID", "INVALID", "DELETED" ]
        },
        "feedId" : {
          "type" : "string"
        },
        "feedName" : {
          "type" : "string"
        },
        "feedVersion" : {
          "type" : "string"
        },
        "feedDescription" : {
          "type" : "string"
        },
        "owner" : {
          "type" : "string"
        },
        "asprClassification" : {
          "type" : "string"
        },
        "publishURL" : {
          "type" : "string"
        },
        "subscribeURL" : {
          "type" : "string"
        },
        "suspended" : {
          "type" : "boolean"
        },
        "logURL" : {
          "type" : "string"
        },
        "formatUuid" : {
          "type" : "string"
        },
        "pubs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DR_Pub"
          }
        },
        "subs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DR_Sub"
          }
        },
        "bytes" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "byte"
          }
        }
      }
    },
...
  }
}